+++ assets/js/warehouse.js
@@ -483,14 +483,6 @@ var ajaxPopState = function() {
 // onunload: Fix for Firefox to execute Javascript on history back.
 window.onunload = function() {};
 
-// Check if logged in.
-// http://stackoverflow.com/questions/6359327/detect-back-button-click-in-browser
-if (window.performance && window.performance.navigation.type == 2) {
-	if (document.URL.indexOf('Modules.php?') != -1) {
-		window.location.href = 'index.php?modfunc=logout';
-	}
-}
-
 // ListOutput JS.
 var LOSearch = function(ev, val, url) {
 	if (ev.type === 'click' || ev.keyCode == 13) {